+++ test/pageobjects/Profile.page.js
@@ -18,61 +18,12 @@ class ProfilePage extends Page {
         return $("//div[@class='profile-image initials']");
     }
 
+    get profileName() {
+        return $("//div[@class = 'MuiTypography-root MuiTypography-h4 css-1xvinid']");
+    }
+
     get profileData () {
         return $(".ant-col.p-3");
-    }
-
-    get inputFirstName () {
-        return $('#first-name');
-    }
-
-    get inputLastName () {
-        return $('#last-name');
-    }
-
-    get inputJobTitle () {
-        return $('#job-title');
-    }
-
-    get inputImageLink () {
-        return $('#image');
-    }
-
-    get inputAbout () {
-        return $('#about');
-    }
-
-    get inputLanguages () {
-        return $('#languages');
-    }
-    get btnSave () {
-        return $('button[type="submit"]');
-    }
-
-    get btnCancel () {
-        return $('button[type="type"]');
-    }
-
-    async fillForm (firstName, lastName, jobTitle, imageLink, about, lang) {
-        await this.open()
-        await this.inputFirstName.setValue(firstName);
-        await this.inputLastName.setValue(lastName);
-        await this.inputJobTitle.setValue(jobTitle);
-        await this.inputImageLink.setValue(imageLink);
-        await this.inputAbout.setValue(about);
-        await this.inputLanguages.setValue(lang);
-    }
-
-    clickBackBtn () {
-        this.btnBack.click();
-    }
-
-    clickSaveBtn () {
-        this.btnSave.click();
-    }
-
-    clickCancelBtn () {
-        this.btnCancel.click();
     }
 
     open() {
